Revert "utils file for func used everywhere in handler"
This reverts commit c008963abcc2e3d9b763cb3482074b04b5091d68.
